@@ -17,7 +17,7 @@
   "dependencies": {
     "@headlessui/react": "^1.7.17",
     "@next-auth/prisma-adapter": "^1.0.7",
-    "@prisma/client": "^5.4.2",
+    "@prisma/client": "^5.19.1",
     "@radix-ui/react-switch": "^1.0.3",
     "@radix-ui/react-toast": "^1.1.4",
     "@radix-ui/react-tooltip": "^1.0.6",
@@ -31,17 +31,17 @@
     "@upstash/ratelimit": "^0.4.4",
     "@vercel/analytics": "^1.0.2",
     "@vercel/edge": "^1.0.1",
-    "axios": "^0.26.0",
+    "axios": "^1.7.7",
     "cheerio": "^1.0.0-rc.12",
     "chromadb": "^1.5.9",
     "clsx": "^2.0.0",
     "framer-motion": "^10.16.4",
     "html-to-image": "^1.11.11",
     "i18next": "^22.5.1",
-    "langchain": "^0.0.91",
+    "langchain": "^0.3.2",
     "micro": "^10.0.1",
     "micro-cors": "^0.1.1",
-    "next": "13.5.5",
+    "next": "^14.2.13",
     "next-auth": "^4.23.2",
     "next-i18next": "^13.3.0",
     "nextjs-google-analytics": "^2.3.3",
@@ -79,7 +79,7 @@
     "husky": "^8.0.3",
     "jest": "^29.6.4",
     "jest-environment-jsdom": "^29.6.3",
-    "lint-staged": "^14.0.0",
+    "lint-staged": "^15.2.10",
     "postcss": "^8.4.30",
     "prettier": "^3.0.3",
     "prettier-plugin-tailwindcss": "^0.5.4",
